+++ tools/spec-gen-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release
 
+## 2025-01-07 - 0.1.2
+
+- Cleaned up dependencies
+- Set variables for pipeline scenarios
+- Added alias for CLI parameters
+
 ## 2024-12-19 - 0.1.1
 
 - Added saveFilterLog function to save the filtered log